manually corrected "Resampling" entries in IDs 11 and 31
led to errors in interpretation (0,3 and 3,9 were saved as 0.3 and 3. Sept). The user input was ambiguous as two numbers were stated for one study. The assumption is that both numbers were stated for both studies in each case.
Assuming for instance 0=no resampling in study and 3 = 3Hz resampling in study 2